+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `AsyncValuesUpdated` event which is the async equivalent of `ValuesUpdated`. (#862)
 - Added `#TIMEOUT!` error returned when async function takes too long to resolve. (#862)
 - Added `timeoutTime` configuration option to customize async functions timeout. (#862)
-- Added `Loading...` text to the translations for async cells with a pending promise. (#862)
+- Added `Loading...` cell error to the translations for async cells with a pending promise. All dependent cells will have this error. (#862)
 
 ## [1.3.0] - 2021-10-20